Add string hashing and RabinKarp algorithms

+++ src/test/java/me/beresnev/algorithms/BinaryPowTest.java
@@ -16,7 +16,7 @@ public class BinaryPowTest {
         Random random = new Random();
         for (int i = 2; i < 50; i++) {
             int pow = random.nextInt(5);
-            int myPow = BinaryPow.pow(i, pow);
+            long myPow = BinaryPow.pow(i, pow);
             double mathPow = Math.pow((double) i, (double) pow);
             Assert.assertEquals(myPow, (int) mathPow);
         }
@@ -24,19 +24,19 @@ public class BinaryPowTest {
 
     @Test
     public void negativePow() {
-        int pow = BinaryPow.pow(5, -3);
+        long pow = BinaryPow.pow(5, -3);
         Assert.assertEquals(pow, 0);
     }
 
     @Test
     public void negativeNumber() {
-        int pow = BinaryPow.pow(-5, 3);
+        long pow = BinaryPow.pow(-5, 3);
         Assert.assertEquals(pow, -125);
     }
 
     @Test
     public void powEqualsZero() {
-        int pow = BinaryPow.pow(5, 0);
+        long pow = BinaryPow.pow(5, 0);
         Assert.assertEquals(pow, 1);
     }
 }
